Skiller alt som har med det nye brevmottaker-konseptet og det gamle verge-konseptet fra hverandre. Oppretter et helt separat behandlingssteg for brevmottakere, men viser det aldri samtidig med verge steget
