+++ src/frontend/context/BehandlingContext.tsx
@@ -31,6 +31,7 @@ const [BehandlingProvider, useBehandling] = createUseContext(() => {
     const [aktivtSteg, settAktivtSteg] = useState<IBehandlingsstegstilstand>();
     const [ventegrunn, settVentegrunn] = useState<IBehandlingsstegstilstand>();
     const [visVenteModal, settVisVenteModal] = useState<boolean>(false);
+    const [visBrevmottakerModal, settVisBrevmottakerModal] = useState<boolean>(false);
     const [harKravgrunnlag, settHarKravgrunnlag] = useState<boolean>();
     const [behandlingILesemodus, settBehandlingILesemodus] = useState<boolean>();
     const [åpenHøyremeny, settÅpenHøyremeny] = useState(true);
@@ -51,7 +52,8 @@ const [BehandlingProvider, useBehandling] = createUseContext(() => {
 
     const hentBehandlingMedBehandlingId = (
         behandlingId: string,
-        henterEtterStatusendring?: boolean | false
+        henterEtterStatusendring?: boolean | false,
+        ønsketPathEtterHenting?: string | undefined
     ): void => {
         settBehandling(byggHenterRessurs());
         settAktivtSteg(undefined);
@@ -109,6 +111,8 @@ const [BehandlingProvider, useBehandling] = createUseContext(() => {
                             // @ts-ignore - fagsak er hentet på dette tidspunktet
                             `/fagsystem/${fagsak?.data?.fagsystem}/fagsak/${fagsak?.data?.eksternFagsakId}/behandling/${hentetBehandling.data.eksternBrukId}`
                         );
+                    } else if (ønsketPathEtterHenting !== undefined) {
+                        navigate(ønsketPathEtterHenting);
                     }
                 }
                 settBehandling(hentetBehandling);
@@ -188,6 +192,8 @@ const [BehandlingProvider, useBehandling] = createUseContext(() => {
         lagLenkeTilRevurdering,
         åpenHøyremeny,
         settÅpenHøyremeny,
+        visBrevmottakerModal,
+        settVisBrevmottakerModal,
     };
 });
