+++ src/frontend/komponenter/Fagsak/BehandlingContainer.tsx
@@ -15,6 +15,7 @@ import {
 } from '../Felleskomponenter/Venstremeny/sider';
 import Venstremeny from '../Felleskomponenter/Venstremeny/Venstremeny';
 import BrevmottakerContainer from './Brevmottaker/BrevmottakerContainer';
+import { BrevmottakerProvider } from './Brevmottaker/BrevmottakerContext';
 import FaktaContainer from './Fakta/FaktaContainer';
 import { FeilutbetalingFaktaProvider } from './Fakta/FeilutbetalingFaktaContext';
 import { FeilutbetalingForeldelseProvider } from './Foreldelse/FeilutbetalingForeldelseContext';
@@ -22,7 +23,8 @@ import ForeldelseContainer from './Foreldelse/ForeldelseContainer';
 import Høyremeny from './Høyremeny/Høyremeny';
 import { FeilutbetalingVedtakProvider } from './Vedtak/FeilutbetalingVedtakContext';
 import VedtakContainer from './Vedtak/VedtakContainer';
-import { BrevmottakerProvider } from './Verge/VergeContext';
+import VergeContainer from './Verge/VergeContainer';
+import { VergeProvider } from './Verge/VergeContext';
 import { FeilutbetalingVilkårsvurderingProvider } from './Vilkårsvurdering/FeilutbetalingVilkårsvurderingContext';
 import VilkårsvurderingContainer from './Vilkårsvurdering/VilkårsvurderingContainer';
 
@@ -157,6 +159,15 @@ const BehandlingContainer: React.FC<IProps> = ({ fagsak, behandling }) => {
 
                     <Route
                         path={BEHANDLING_KONTEKST_PATH + '/verge'}
+                        element={
+                            <VergeProvider behandling={behandling} fagsak={fagsak}>
+                                <VergeContainer />
+                            </VergeProvider>
+                        }
+                    />
+
+                    <Route
+                        path={BEHANDLING_KONTEKST_PATH + '/brevmottakere'}
                         element={
                             <BrevmottakerProvider behandling={behandling} fagsak={fagsak}>
                                 <BrevmottakerContainer />
